mail: split into a core that does not draw and a graphical front end
The mail program is now two files. portable/mailcore.c is the program
proper -- the store, the index, the digests, the accounts, the reading
of messages, the talking to servers, and the thread that does all of it
-- and it draws nothing: it compiles without graphics.h and touches
nothing of the library but services and network. graph_programs/mail.c
is what is left, which is everything that puts it on a screen.

mailcore.h states the whole contract between them. A front end supplies
three functions, since only it knows how to say anything -- status(),
fail(), statprog() -- and keeps one rule: hold the lock for as long as
it is handling an event. The core takes the same lock around each
message it commits and around anything that rebuilds the folder table,
and nowhere else.

The cut was made by moving lines, not regenerating them: every section
heading and every comment came across whole. Most of the program was
already indifferent to how it is displayed, which is why the core came
out at 3,900 lines against 2,600 of display.

This is the ground for mailc, a character front end over the same core,
which is the point of the exercise: the core cannot tell which one is
running it.

Proven against the whole battery: a fetch with the window resized and
moved throughout (200 stored, 20 deduplicated), a restart reading the
index back in 1ms with dedup holding, a send with its local Sent copy,
and the window pixel for pixel as it was.
